Use Express
In general, for a RESTful API project I would prefer Restify. It allows
to do [API versioning][1] relatively easy and has to offer other nice
and well-tailored features.

Since Restify is not on our list we have to choose between hapi.js and
Express.

hapi is a very good and mature framework proven to be secure, stable
and performant enough, but for this particular use case I decided to go
with a lightweight Express.

Generic [performance benchmarks][2] demonstrate that Express tends to
be faster than hapi, but those tests are [doubtful][3] and a subject to
debate as we choose frameworks based not only on their ability to
process bare requests, but also on features they have (or don't) and
how they match our priorities.

Also, Fastify is worth mentioning. It's a new framework that has been
recently [introduced][4] and already deserves some attention as it
looks really promising.

[1]: https://stripe.com/blog/api-versioning
[2]: https://raygun.com/blog/node-js-performance-2017/
[3]: https://hueniverse.com/performance-at-rest-75bb8fff143
[4]: https://thenewstack.io/introducing-fastify-speedy-node-js-web-framework/