Pick a package to work with polygons
I went to npms.io in order to check out which packages are currently
popular.

At a glance Turf module seems to be one of the most used packages when
it comes to geographic algorithms. For example, `@turf/within` is a
function that finds all points that fall within specified polygons. I
decided to take a look at its [implementation][1] first as it is a good
starting point before actually opting for it or anything else.

`@turf/within` uses underlying `@turf/inside` function and has the most
straightforward implementation time complexity of which is O(n^2).

Turf also works only with feature objects which means all primitives
including a simple point must be wrapped into a feature object before
Turf can do anything with it.

That made me think that there could be another alternatives.

I searched for other popular libraries and made a short list of these:
- [`@turf/inside`](https://www.npmjs.com/package/%40turf%2Finside)
- [`point-in-polygon`](https://www.npmjs.com/package/point-in-polygon)
- [`polygon-lookup`](https://www.npmjs.com/package/polygon-lookup)

After reading documentation for all of them `polygon-lookup` which is
built on top of `point-in-polygon` looked the most promising. It uses
tree data structures to narrow down candidate polygons for a point.
Source code was confirming my feelings, but I decided to go for a quick
benchmark just to make sure that it really performs fast enough.

`polygon-lookup` has actually demonstrated pretty high performance
which doesn't really depend on input size.

So I decided to go with it for polygon lookup for a point and with
`point-in-polygon` for filtering points that are inside of a polygon.

For numbers, please, run `npm run benchmark`.

[1]: https://github.com/Turfjs/turf/blob/v4.7.2/packages/turf-within/index.js#L40-L51