Simple HTTP GET client
We will need to query upstream API, so we need a convenient way of
doing that.

Even though one of the greatest benefits of Node.js and NPM ecosystem
is modules that enable developers to build their project faster, I
decided to opt out of the popular `request` module, for example, and to
go with own tiny cholesterol-free module.

This project is not request-intensive, so there's no need in carrying
extra code for handling redirects, for example, or other request
methods other than this simple GET over HTTPS.

Anyways, in case of project growth it can be easily replaced in future
with one of those great and ready-to-use packages.